refactor: Refactors user module imports and centralizes exports
Unifies and streamlines import paths for user-related modules by
centralizing exports through index files. Updates constructor property
naming for consistency and improves overall module organization, making
future maintenance and scaling easier.
