+++ src/container/user.container.ts
@@ -1,8 +1,8 @@
 //Padrão utilizado para buildar o fluxo de maneira profissional
 //Para chamada no routes ficar limpa
 
-import { UserRepository } from './../database/user.repository';
-import { UserController } from "../controllers/user.controller";
+import { UserRepository } from './../database';
+import { UserController } from "../controllers";
 import { UserService } from "../services";
 
 const userRepository = new UserRepository();
