+++ src/controllers/index.ts
@@ -1,1 +1,2 @@
 export * from "./examples.controller";
+export * from "./user.controller";
