+++ src/database/index.ts
@@ -1,1 +1,2 @@
 export * from "./prisma.repository";
+export * from "./user.repository";
