+++ src/routes/index.ts
@@ -1,1 +1,2 @@
 export * from "./example.routes";
+export * from "./users.routes";
