+++ src/server.ts
@@ -2,9 +2,9 @@ import { log } from "console";
 import App from "./app";
 import { envs } from "./envs";
 import {
-  ExampleRoutes
+  ExampleRoutes,
+  UsersRoutes
 } from "./routes";
-import { UsersRoutes } from "./routes/users.routes";
 
 
 if (envs.PORT === undefined || envs.PORT == null) {
@@ -14,9 +14,7 @@ if (envs.PORT === undefined || envs.PORT == null) {
 const app = new App(
   [
     ExampleRoutes.bind(),
-    UsersRoutes.bind()
-    // Add more routes here
-
+    UsersRoutes.bind(),
   ],
   envs.PORT,
 );
